requirements.txt

Flask==2.3.3
Werkzeug==2.3.7

SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
greenlet==3.0.0

Flask-JWT-Extended==4.5.2
PyJWT==2.8.0

Flask-Cors==4.0.0


marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

python-dotenv==1.0.0
pydantic==2.4.2
pytz==2023.3
click==8.1.7